third_party/blink/web_tests/fast/css/add-remove-stylesheets-minimal-recalc-style-expected.txt

PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS successfullyParsed is true

TEST COMPLETE

foo
foo
foo
bar
bar
bar
baz
baz
baz
